refactor(insights): move TakeawayCard to shared tabs/components and drop per-tab viz dirs
Move TakeawayCard from tabs/engagement/viz/ to the shared tabs/components/
directory, following the same pattern used for the other hoisted viz
components (Funnel, LineChart, PieChart, BarChart, SortableTable,
DistributionTable). Update ReaderSegmentsSection to import from the new
shared location and delete the now-empty engagement/viz/ directory.

Also: remove the stale "PieChart / LineChart / SortableTable are candidates
to hoist" comment from conversion.scss (all three are already hoisted), and
add pct assertions to the DistributionTable test so the row-render test
actually covers the percent column it claims to cover.
